fab_and_fit/requirements.txt
django 
djangorestframework 
google-genai 
Pillow 
python-dotenv 
dotenv 
google-generativeai 
django-crispy-forms 
genai
djangorestframework 
djangorestframework-simplejwt